Change field resolvers to be real GraphQL resolvers
The final code is much cleaner and following known GraphQL patterns now :)
