+++ graphql/index.ts
@@ -1,5 +1,9 @@
 import './axiosConfig'
 import catalogQueries from './resolvers/catalog'
+import Brand from './resolvers/catalog/Brand'
+import Category from './resolvers/catalog/Category'
+import Facets from './resolvers/catalog/Facets'
+import Product from './resolvers/catalog/Product'
 import {mutations as checkoutMutations, queries as checkoutQueries} from './resolvers/checkout'
 import {mutations as profileMutations, queries as profileQueries} from './resolvers/profile'
 
@@ -17,4 +21,8 @@ export const resolvers = {
     ...profileQueries,
     ...checkoutQueries,
   },
+  ...Brand,
+  ...Category,
+  ...Facets,
+  ...Product,
 }
