+++ graphql/package.json
@@ -6,7 +6,6 @@
     "bluebird": "^3.5.0",
     "co-body": "^5.1.1",
     "cookie": "^0.3.1",
-    "graphql-fields": "^1.0.2",
     "ramda": "^0.24.1",
     "slugify": "^1.2.6",
     "ts-node": "3.3.0",
