+++ graphql/resolvers/catalog/Brand.ts
@@ -1,0 +1,8 @@
+import * as slugify from 'slugify'
+
+export default {
+  Brand: {
+    active: ({isActive}) => isActive,
+    slug: ({name}) => slugify(name, {lower: true}),
+  },
+}
