+++ graphql/resolvers/catalog/Category.ts
@@ -1,0 +1,10 @@
+import {compose, last, split} from 'ramda'
+
+const getLastPathSegment = compose(last, split('/'))
+
+export default {
+  Category: {
+    href: ({url}) => url,
+    slug: ({url}) => url ? getLastPathSegment(url) : null,
+  },
+}
