+++ graphql/resolvers/catalog/Facets.ts
@@ -1,0 +1,10 @@
+import {objToNameValue} from './objToNameValue'
+
+export default {
+  Facets: {
+    SpecificationFilters: (facets) => {
+      const {SpecificationFilters={}} = facets
+      return objToNameValue('name', 'value', SpecificationFilters)
+    }
+  },
+}
