+++ graphql/resolvers/catalog/index.ts
@@ -1,43 +1,33 @@
-import axios, {AxiosResponse} from 'axios'
+import axios from 'axios'
 import {ColossusContext} from 'colossus'
-import graphqlFields from 'graphql-fields'
-import {compose, equals, find, map, prop} from 'ramda'
+import {compose, equals, find, prop} from 'ramda'
 import ResolverError from '../../errors/resolverError'
 import {withAuthToken} from '../headers'
 import paths from '../paths'
-import {resolveBrandFields, resolveCategoryFields, resolveFacetFields, resolveProductFields} from './fieldsResolver'
 
 export default {
   autocomplete: async (_, data, {vtex: ioContext}: ColossusContext) => {
     const url = paths.autocomplete(ioContext.account, data)
     const {data: resolvedAutocomplete} = await axios.get(url, { headers: withAuthToken()(ioContext) })
-
     return resolvedAutocomplete
   },
 
   facets: async (_, data, {vtex: ioContext}: ColossusContext) => {
     const url = paths.facets(ioContext.account, data)
     const {data: facets} = await axios.get(url, { headers: withAuthToken()(ioContext) })
-    const resolvedFacets = resolveFacetFields(facets)
-
-    return resolvedFacets
+    return facets
   },
 
-  product: async (_, data, {vtex: ioContext}: ColossusContext, info) => {
+  product: async (_, data, {vtex: ioContext}: ColossusContext) => {
     const url = paths.product(ioContext.account, data)
     const {data: product} = await axios.get(url, { headers: withAuthToken()(ioContext) })
-    const resolvedProduct = await resolveProductFields(ioContext, product, graphqlFields(info))
-
-    return resolvedProduct
+    return product
   },
 
-  products: async (_, data, {vtex: ioContext}: ColossusContext, info) => {
+  products: async (_, data, {vtex: ioContext}: ColossusContext) => {
     const url = paths.products(ioContext.account, data)
     const {data: products} = await axios.get(url, { headers: withAuthToken()(ioContext) })
-    const fields = graphqlFields(info)
-    const resolvedProducts = await Promise.map(products, product => resolveProductFields(ioContext, product, fields))
-
-    return resolvedProducts
+    return products
   },
 
   brand: async (_, data, {vtex: ioContext, request: {headers: {cookie}}}: ColossusContext) => {
@@ -48,18 +38,18 @@ export default {
     if (!brand) {
       throw new ResolverError(`Brand with id ${data.id} not found`, 404)
     }
-    return resolveBrandFields(brand)
+    return brand
   },
 
   category: async (_, data, {vtex: ioContext, request: {headers: {cookie}}}: ColossusContext) => {
     const url = paths.category(ioContext.account, data)
     const {data: category} = await axios.get(url, {headers: withAuthToken()(ioContext, cookie) })
-    return resolveCategoryFields(category)
+    return category
   },
 
   categories: async (_, data, {vtex: ioContext}: ColossusContext) => {
     const url = paths.categories(ioContext.account, data)
     const {data: categories} = await axios.get(url, {headers: withAuthToken()(ioContext) })
-    return map(resolveCategoryFields, categories)
+    return categories
   }
 }
